+++ pkg/store/bucket.go
@@ -120,6 +120,7 @@ type bucketStoreMetrics struct {
 	chunkSizeBytes        prometheus.Histogram
 	queriesDropped        *prometheus.CounterVec
 	seriesRefetches       prometheus.Counter
+	emptyPostingCount     prometheus.Counter
 
 	cachedPostingsCompressions           *prometheus.CounterVec
 	cachedPostingsCompressionErrors      *prometheus.CounterVec
@@ -253,6 +254,11 @@ func newBucketStoreMetrics(reg prometheus.Registerer) *bucketStoreMetrics {
 		Name:    "thanos_bucket_store_cached_postings_fetch_duration_seconds",
 		Help:    "The time it takes to fetch postings to respond to a request sent to a store gateway. It includes both the time to fetch it from the cache and from storage in case of cache misses.",
 		Buckets: []float64{0.001, 0.01, 0.1, 0.3, 0.6, 1, 3, 6, 9, 20, 30, 60, 90, 120},
+	})
+
+	m.emptyPostingCount = promauto.With(reg).NewCounter(prometheus.CounterOpts{
+		Name: "thanos_bucket_store_empty_postings_total",
+		Help: "Total number of empty postings when fetching block series.",
 	})
 
 	return &m
@@ -789,6 +795,7 @@ func blockSeries(
 	minTime, maxTime int64, // Series must have data in this time range to be returned.
 	loadAggregates []storepb.Aggr, // List of aggregates to load when loading chunks.
 	shardMatcher *storepb.ShardMatcher,
+	emptyPostingsCount prometheus.Counter,
 ) (storepb.SeriesSet, *queryStats, error) {
 	ps, err := indexr.ExpandedPostings(ctx, matchers)
 	if err != nil {
@@ -796,6 +803,7 @@ func blockSeries(
 	}
 
 	if len(ps) == 0 {
+		emptyPostingsCount.Inc()
 		return storepb.EmptySeriesSet(), indexr.stats, nil
 	}
 
@@ -1086,6 +1094,7 @@ func (s *BucketStore) Series(req *storepb.SeriesRequest, srv storepb.Store_Serie
 					req.MinTime, req.MaxTime,
 					req.Aggregates,
 					shardMatcher,
+					s.metrics.emptyPostingCount,
 				)
 				if err != nil {
 					return errors.Wrapf(err, "fetch series for block %s", b.meta.ULID)
@@ -1293,7 +1302,21 @@ func (s *BucketStore) LabelNames(ctx context.Context, req *storepb.LabelNamesReq
 
 				result = strutil.MergeSlices(res, extRes)
 			} else {
-				seriesSet, _, err := blockSeries(newCtx, b.extLset, indexr, nil, reqSeriesMatchersNoExtLabels, nil, seriesLimiter, true, req.Start, req.End, nil, nil)
+				seriesSet, _, err := blockSeries(
+					newCtx,
+					b.extLset,
+					indexr,
+					nil,
+					reqSeriesMatchersNoExtLabels,
+					nil,
+					seriesLimiter,
+					true,
+					req.Start,
+					req.End,
+					nil,
+					nil,
+					s.metrics.emptyPostingCount,
+				)
 				if err != nil {
 					return errors.Wrapf(err, "fetch series for block %s", b.meta.ULID)
 				}
@@ -1447,7 +1470,21 @@ func (s *BucketStore) LabelValues(ctx context.Context, req *storepb.LabelValuesR
 				}
 				result = res
 			} else {
-				seriesSet, _, err := blockSeries(newCtx, b.extLset, indexr, nil, reqSeriesMatchersNoExtLabels, nil, seriesLimiter, true, req.Start, req.End, nil, nil)
+				seriesSet, _, err := blockSeries(
+					newCtx,
+					b.extLset,
+					indexr,
+					nil,
+					reqSeriesMatchersNoExtLabels,
+					nil,
+					seriesLimiter,
+					true,
+					req.Start,
+					req.End,
+					nil,
+					nil,
+					s.metrics.emptyPostingCount,
+				)
 				if err != nil {
 					return errors.Wrapf(err, "fetch series for block %s", b.meta.ULID)
 				}